+++ dcb/src/test/java/org/olf/dcb/ingest/FolioIngestTests.java
@@ -17,6 +17,7 @@ import static org.olf.dcb.test.matchers.BibRecordMatchers.hasTitleMetadata;
 import java.util.List;
 
 import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
 import org.mockserver.client.MockServerClient;
 import org.olf.dcb.core.model.BibRecord;